Relocate internal Sphinx extension
Move it up from `pfsc.sphinx.sphinx_proofscape` to just `pfsc.sphinx`.
