+++ server/req/requirements.in
@@ -55,7 +55,7 @@ mistletoe==0.7.2
 furo==2022.12.7
 # The *external* `sphinx-proofscape` package provides syntax highlighting for
 # the pfsc module syntax and meson language, while the Sphinx extension defined
-# locally in `pise/server` under `pfsc.sphinx.sphinx_proofscape` defines
+# locally in `pise/server` under `pfsc.sphinx` defines
 # everything else (roles and directives for widgets, etc.).
 sphinx-proofscape==0.4.0
 # When you write a Sphinx doc inside a Proofscape repo, you can use dollar
